Make sure just uploaded files can directly be edited from "BP Media Editor"
As the "BackBone" Uploader is creating Attachments using the WordPress Attachment model, we need to override 2 functions used in wp-plupload.js :
- media.model.Attachment.create
- media.model.Attachment.get

Doing so, the user is able to edit the title and the description of the different files uploaded at any time directly from the "BP Media Uploader". He can also delete the file or open the file edit screen.

Fixes #1