Main navigation when Groups component is active
Adapt the REST Controller so that it is possible for the member to view the content of the `My Groups` and `My Media` primary directories.
Adapt the `bp_attachments_list_member_root_objects()` function to fetch the directory types when the user requests his media. Listing the Group folders is still needed.Avoid displaying the folder icon at 100% when it is an avatar
Update WPCLI and WPCS