Improve the Directory Tree & current Directory management
Use the title as the name and the name as the slug when setting a Directory tree entry: this improves the TreeSelect control appearance in particular it uses translated strings for initial directories.
Add a new property to the directory item. The `readonly` property makes it possible to only show the "Add new" button when it is possible to upload a file or create a subdirectory.
This commit also adds two new SVGs to replace the user avatar for the "My Uploads" directory